refactor: move `AvatarUploader` to the `widgets` folder
Renamed `UseAvatarUploader` to `use-avatar-uploader`.
