--- features/avatar-uploader/index.ts
@@ -1,3 +1,0 @@
-export * from './avatar-uploader';
-export * from './avatar-uploader.stories';
-export * from './avatar-uploader.module.scss';
