+++ widgets/avatar-uploader/avatar-uploader.module.scss
