+++ widgets/avatar-uploader/index.ts
@@ -1,0 +1,1 @@
+export * from './avatar-uploader';
